+++ src/platform/site-wide/va-footer/helpers.jsx
@@ -1,0 +1,34 @@
+import React from 'react';
+import orderBy from 'lodash/orderBy';
+
+import recordEvent from '../../monitoring/record-event';
+
+export const FOOTER_COLUMNS = {
+  PROGRAMS: '1',
+  RESOURCES: '2',
+  CONNECT: '3',
+  CONTACT: '4',
+};
+
+export const FOOTER_EVENTS = {
+  [FOOTER_COLUMNS.PROGRAMS]: 'nav-footer-programs',
+  [FOOTER_COLUMNS.RESOURCES]: 'nav-footer-resources',
+  [FOOTER_COLUMNS.CONNECT]: 'nav-footer-connect',
+  [FOOTER_COLUMNS.CONTACT]: 'nav-footer-contact',
+  CRISIS_LINE: 'nav-footer-crisis',
+};
+
+export function generateLinkItems(links, column, direction = 'asc') {
+  const captureEvent = () => recordEvent({ event: FOOTER_EVENTS[column] });
+  return (
+    <ul className="va-footer-links">
+      {orderBy(links[column], 'order', direction).map(link => (
+        <li key={`${link.column}-${link.order}`}>
+          <a href={link.href} onClick={captureEvent} target={link.target}>
+            {link.title}
+          </a>
+        </li>
+      ))}
+    </ul>
+  );
+}
